# frontend/test/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis",
    "description": "Static-scan hits + LSP call graph -> AI-ready analysis bundles. First deliverable: the assembly stage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/audit": {
      "post": {
        "summary": "Submit an autonomous repository audit (the agent harness, as a queued job)",
        "operationId": "submit_audit_v1_audit_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Run it again even though a previous attempt finished",
              "default": false,
              "title": "Force"
            },
            "description": "Run it again even though a previous attempt finished"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/{job_id}/trail": {
      "get": {
        "summary": "Audit Trail",
        "description": "The run's events after `after_seq`, oldest first.\n\n`exists: false` with an empty list means \"this audit has not written a trail yet\" -- a queued\njob, or one that has not reached its first stage. `more: true` means the page filled up and the\ncaller should immediately ask again rather than wait for the next poll tick.\n\n`job.running` is included because the console needs to know whether to keep polling, and\nderiving it from `events[-1].kind == \"summary\"` would make a killed run look finished.",
        "operationId": "audit_trail_v1_audit__job_id__trail_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Return only events with a larger seq",
              "default": 0,
              "title": "After Seq"
            },
            "description": "Return only events with a larger seq"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/{job_id}/report": {
      "get": {
        "summary": "Audit Report",
        "description": "The finished markdown report, or a 404 while the run is still going.\n\nReturned as JSON with the text in a field rather than as `text/markdown`, because the console\nrenders it in a pane with other metadata and a bare text body would leave it guessing which\nfields exist. The run directory is included so a reader can find the blackboard and the trail\nnext to it.",
        "operationId": "audit_report_v1_audit__job_id__report_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuditRequest": {
        "properties": {
          "workspace": {
            "type": "string",
            "title": "Workspace",
            "description": "Repo path inside the container/workspace, e.g. /data/projects/x."
          }
        },
        "type": "object",
        "required": [
          "workspace"
        ],
        "title": "AuditRequest",
        "description": "Ask the agent harness to review a whole repository.\n\nOnly `workspace` is required, and only `workspace` is accepted: the harness bounds\n(`max_rounds`, `steps_per_agent`, concurrency) are deliberately *not* per-request fields.\nThey are part of what the review produces, so letting a caller vary them would make two\nrequests for the same repository look identical while producing different findings -- and the\njob fingerprint, which is how a resubmission attaches instead of running twice, is built from\nexactly the fields here."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}